적용 완료
