+++ src/component/TodoList.jsx
@@ -1,19 +1,22 @@
-import React, { memo } from "react";
+import React, { memo, useCallback } from "react";
 import TodoDateRow from "./TodoDateRow";
 import Todo from "./Todo";
+import {compare} from './fs';
 
 import { useDispatch, useSelector } from 'react-redux';
-import { setTodoDatum } from '../redux/reducers/memoSlice';
+import { setDoneDatum } from '../redux/reducers/memoSlice';
+import {saveData} from './fs';
 
 const TodoList = (props) => {
+  const { onDeleteTodo } = props;
+
   const todoDatum = useSelector(state => state.memo.todoDatum);
-  // console.log(todoDatum);
-  const { onDeleteTodo, onDoneTodo} = props;
+  const doneDatum = useSelector(state => state.memo.doneDatum);
+  
+  const dispatch = useDispatch();
 
   const onClickRouter = ({ target }) => {
     const { id } = target.parentNode;
-    
-    // console.log('id', id);
 
     if (target.matches("#delete")) {
       handleDeleteTodo(id);
@@ -21,12 +24,30 @@ const TodoList = (props) => {
       handleDoneTodo(id);
     }
   };
-  const handleDeleteTodo = (id) => {
+
+  const handleDeleteTodo = useCallback((id) => {
     onDeleteTodo(id, todoDatum, "todoDatum");
-  };
+  },[onDeleteTodo, todoDatum]);
+
+  const onDoneTodo = useCallback(
+    (key, todoDatum) => {
+      let doneTodo = handleDeleteTodo(key, todoDatum, "todoDatum"); // [{}]
+
+      doneTodo = todoDatum.filter((a) => a.startDate === Number(key));
+
+      doneTodo.doneDate = Date.now();
+      const newDoneTodo = [...doneTodo, ...doneDatum].sort(compare("doneDate"));
+      
+      dispatch(setDoneDatum(newDoneTodo));
+      saveData("doneDatum", newDoneTodo);
+    },
+    [handleDeleteTodo, doneDatum, dispatch]
+  );
+
   const handleDoneTodo = (id) => {
     onDoneTodo(id, todoDatum);
   };
+
   const copyDatum = [...todoDatum];
   const rows = [];
   let lastCategory = null;
@@ -65,4 +86,4 @@ const TodoList = (props) => {
   );
 };
 
-export default (TodoList);
+export default memo(TodoList);
