+++ src/component/fs.js
@@ -1,20 +1,5 @@
-// Container
 export const compare = (key) => (a, b) =>
   a[key] < b[key] ? 1 : a[key] > b[key] ? -1 : 0;
-
-export const filter = (f, iter) => {
-  const arr = [];
-  for (const a of iter) {
-    if (f(a)) {
-      arr.push(a);
-    }
-  }
-  return arr;
-};
-// export const parseAndSet = (setFunc, dataName) => {
-//   const pasredTodoDatum = JSON.parse(dataName);
-//   setFunc(pasredTodoDatum);
-// };
 
 // CompletedList, todoDateRow
 export const makeDate = (date) => {
@@ -27,3 +12,7 @@ export const makeDate = (date) => {
 };
 
 export const log = console.log;
+
+export const saveData = (name, todoDatum) => {
+  localStorage.setItem(name, JSON.stringify(todoDatum));
+};
