--- src/redux/actions/memoActio.js
@@ -1,10 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-
-export const getMemos = createAsyncThunk(
-  'memo/fetchMemo',
-  async () => {
-    const response =  window.localStorage.getItem('todoDatum');
-    return response;
-  }
-);
-
