--- src/redux/fs.js
@@ -1,3 +1,0 @@
-export const saveData = (name, todoDatum) => {
-  localStorage.setItem(name, JSON.stringify(todoDatum));
-};
